fieldMap: ENH: Use expressions targeting the same fieldNr in sequence instead of reusing the first expression
This is what the user expects: Each AWK-EXPR is applied once, from left to right. This can be implemented by keeping a count of applications for each fieldNr, and passing the corresponding number into evaluateExpression(). Each expression conditional decrements the number; short-circuiting ensures that only those conditionals that match the current fieldNr do this. Only that conditional that arrives at a decremented sequence number of 0 gets executed, so with each invocation, the one further down wins.