fieldGrep: ENH: Allow unbuffered output via -u|--unbuffered
It's a bit more complex here because we have a pipeline of awk ... | grep ... | awk ...
To obtain the default unbuffered output when the command prints to standard out, we the first awk and grep already need arguments for unbuffering. When --unbuffered is explicitly given, all three need that argument.
Introduce grepBufferedArgs and isBuffered, auto-detect the former and use that to also change the first awk behavior.