FIX: The POSIX shell used by AWK does not understand $'\t'
On Ubuntu, the default shell is dash (/bin/sh); it does not support $'...' quoting (unlike Bash and Korn shell). So %q quoting is of no use here.
Instead, reuse my changeQuoting command to do POSIX-compatible (but uglier) quoting. First for the FORMAT itself, then again for rendering the memoizeLines command.