ENH: Support -x|--line-regexp
This cannot be passed directly to grep, because it would prevent the matching of the separator (unfiltered) lines.
For normal pattern matching, all we need to do is add the ^\(...\)$ anchors.
For literal matching, we delegate the escaping to toLiteralRegexp, which conveniently understands all regexp dialects (which are now parsed separately into grepRegexpDialectArgs).