addField: ENH: Allow adding fields at the end with +N syntax
For feature parity with fieldMap, and because here is also is useful to:
- add new fields after the last existing field with empty field(s) in between; previously, only direct appending with -1 was possible
- add multiple new fields after the last existing field; previously, the first expression for -1 won

Because of the different algorithm that maintains the references to the original field numbers via the two-stage evaluation and application, this actually was a lot more difficult than the fieldMap change.
I hope this is a consistent implementation and that the differentiation between -1 and +0 proves useful and understandable.